loading saved list with loader
