+++ app/src/main/java/uco374386/movio2/pv256/fi/muni/cz/filmovarka/Database/MovieDbManager.java
@@ -27,6 +27,7 @@ public class MovieDbManager {
     public static final int COL_MOVIE_POPULARITY = 5;
     public static final int COL_MOVIE_BACKDROP_PATH = 6;
     public static final int COL_MOVIE_COVER_PATH = 7;
+    public static final int COL_MOVIE_IMAGE_BASE = 8;
     private static final String[] MOVIE_COLUMNS = {
             MovieEntry._ID,
             MovieEntry.COLUMN_MOVIEDB_ID,
@@ -36,6 +37,7 @@ public class MovieDbManager {
             MovieEntry.COLUMN_POPULARITY_TEXT,
             MovieEntry.COLUMN_BACKDROP_PATH_TEXT,
             MovieEntry.COLUMN_COVER_PATH_TEXT,
+            MovieEntry.COLUMN_IMAGE_BASE
     };
 
     private static final String WHERE_MOVIEDB_ID = MovieEntry.COLUMN_MOVIEDB_ID + " = ?";
@@ -73,7 +75,9 @@ public class MovieDbManager {
         if (movie.coverPath == null) {
             throw new IllegalStateException("movie coverPath cannot be null");
         }
-
+        if (movie.imageBasePath == null) {
+            throw new IllegalStateException("movie imageBasePath cannot be null");
+        }
         movie.localDbId = ContentUris.parseId(mContext.getContentResolver().insert(MovieEntry.CONTENT_URI, prepareMovieValues(movie)));
     }
 
@@ -102,7 +106,9 @@ public class MovieDbManager {
         if (movie.coverPath == null) {
             throw new IllegalStateException("movie coverPath cannot be null");
         }
-
+        if (movie.imageBasePath == null) {
+            throw new IllegalStateException("movie imageBasePath cannot be null");
+        }
         mContext.getContentResolver().update(MovieEntry.CONTENT_URI, prepareMovieValues(movie), WHERE_MOVIEDB_ID, new String[]{String.valueOf(movie.movieDbId)});
     }
 
@@ -133,6 +139,24 @@ public class MovieDbManager {
         return null;
     }
 
+    public List<MovieResponse> getAll() {
+        Cursor cursor = mContext.getContentResolver().query(MovieEntry.CONTENT_URI, MOVIE_COLUMNS, null, null, null);
+        if (cursor != null && cursor.moveToFirst()) {
+            List<MovieResponse> movies = new ArrayList<>(cursor.getCount());
+            try {
+                while (!cursor.isAfterLast()) {
+                    movies.add(getMovie(cursor));
+                    cursor.moveToNext();
+                }
+            } finally {
+                cursor.close();
+            }
+            return movies;
+        }
+
+        return Collections.emptyList();
+    }
+
     private ContentValues prepareMovieValues(MovieResponse movie) {
         ContentValues values = new ContentValues();
         values.put(MovieEntry.COLUMN_MOVIEDB_ID, movie.movieDbId);
@@ -142,6 +166,7 @@ public class MovieDbManager {
         values.put(MovieEntry.COLUMN_POPULARITY_TEXT, movie.popularity);
         values.put(MovieEntry.COLUMN_BACKDROP_PATH_TEXT, movie.backdropPath);
         values.put(MovieEntry.COLUMN_COVER_PATH_TEXT, movie.coverPath);
+        values.put(MovieEntry.COLUMN_IMAGE_BASE, movie.imageBasePath);
 
         return values;
     }
@@ -156,11 +181,11 @@ public class MovieDbManager {
         } catch (ParseException e) {
             e.printStackTrace();
         }
-        movie.voteAverage = cursor.getDouble(COL_MOVIE_VOTE);
-        movie.popularity = cursor.getDouble(COL_MOVIE_POPULARITY);
+        movie.voteAverage = cursor.getFloat(COL_MOVIE_VOTE);
+        movie.popularity = cursor.getFloat(COL_MOVIE_POPULARITY);
         movie.backdropPath = cursor.getString(COL_MOVIE_BACKDROP_PATH);
         movie.coverPath = cursor.getString(COL_MOVIE_COVER_PATH);
-
+        movie.imageBasePath = cursor.getString(COL_MOVIE_IMAGE_BASE);
         return movie;
     }
