+++ app/src/main/res/layout-sw600dp/content_main.xml
@@ -15,11 +15,11 @@
         android:orientation="horizontal"
         android:layout_height="match_parent">
 
-        <fragment class="uco374386.movio2.pv256.fi.muni.cz.filmovarka.ListFragment"
-                  android:id="@+id/list1"
-                  android:layout_weight="1"
-                  android:layout_width="match_parent"
-                  android:layout_height="match_parent"/>
+        <FrameLayout
+            android:id="@+id/list1"
+            android:layout_weight="1"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"/>
 
         <FrameLayout
             android:id="@+id/details"
