+++ app/src/main/res/layout/content_main.xml
@@ -10,9 +10,10 @@
     tools:context="uco374386.movio2.pv256.fi.muni.cz.filmovarka.MainActivity"
     tools:showIn="@layout/app_bar_main">
 
-    <fragment class="uco374386.movio2.pv256.fi.muni.cz.filmovarka.ListFragment"
-              android:id="@+id/list1"
-              android:layout_width="match_parent"
-              android:layout_height="wrap_content"/>
+    <FrameLayout
+        android:id="@+id/list1"
+        android:layout_weight="1"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"/>
 
 </RelativeLayout>
